for #2556, XXXIntegrateSQLParsingTest => XXXParameterizedSQLParsingTest
