+++ sharding-core/sharding-core-parse/sharding-core-parse-test/src/test/java/org/apache/shardingsphere/core/parse/integrate/engine/AbstractBaseParameterizedParsingTest.java
@@ -39,7 +39,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 @RunWith(Parameterized.class)
-public abstract class AbstractBaseIntegrateSQLParsingTest {
+public abstract class AbstractBaseParameterizedParsingTest {
     
     @Getter(AccessLevel.PROTECTED)
     private static ShardingRule shardingRule;
@@ -54,7 +54,7 @@ public abstract class AbstractBaseIntegrateSQLParsingTest {
     }
     
     private static ShardingRule buildShardingRule() throws IOException {
-        URL url = AbstractBaseIntegrateSQLParsingTest.class.getClassLoader().getResource("yaml/sharding-rule.yaml");
+        URL url = AbstractBaseParameterizedParsingTest.class.getClassLoader().getResource("yaml/sharding-rule.yaml");
         Preconditions.checkNotNull(url, "Cannot find parse rule yaml configuration.");
         YamlRootShardingConfiguration yamlShardingConfig = YamlEngine.unmarshal(new File(url.getFile()), YamlRootShardingConfiguration.class);
         return new ShardingRule(new ShardingRuleConfigurationYamlSwapper().swap(yamlShardingConfig.getShardingRule()), yamlShardingConfig.getDataSources().keySet());
